--- a/openmp/libomptarget/plugins/hsa/impl/atmi_kl.h
+++ b/openmp/libomptarget/plugins/hsa/impl/atmi_kl.h
@@ -36,7 +36,4 @@ typedef struct atmi_kernel_enqueue_template_s {
   void *kernarg_regions;
 } atmi_kernel_enqueue_template_t;
 
-extern void atmid_task_launch(atmi_lparm_t *lp, unsigned long kernel_id,
-                              void *args_region, size_t args_region_size);
-
 #endif  // INCLUDE_ATMI_KL_H_