--- a/openmp/libomptarget/plugins/hsa/impl/internal.h
+++ b/openmp/libomptarget/plugins/hsa/impl/internal.h
@@ -224,10 +224,6 @@ extern hsa_amd_memory_pool_t get_memory_pool_by_mem_place(
     atmi_mem_place_t place);
 extern bool atl_is_atmi_initialized();
 
-// extern void atl_task_wait(TaskImpl *task);
-
-void init_dag_scheduler();
-bool handle_signal(hsa_signal_value_t value, void *arg);
 bool handle_group_signal(hsa_signal_value_t value, void *arg);
 
 void packet_store_release(uint32_t *packet, uint16_t header, uint16_t rest);
@@ -236,14 +232,6 @@ uint16_t create_header(
     atmi_task_fence_scope_t acq_fence = ATMI_FENCE_SCOPE_SYSTEM,
     atmi_task_fence_scope_t rel_fence = ATMI_FENCE_SCOPE_SYSTEM);
 
-Kernel *get_kernel_obj(atmi_kernel_t atmi_kernel);
-TaskImpl *getTaskImpl(atmi_task_handle_t t);
-core::TaskgroupImpl *getTaskgroupImpl(atmi_taskgroup_handle_t t);
-void set_task_handle_ID(atmi_task_handle_t *t, int ID);
-void lock(pthread_mutex_t *m);
-void unlock(pthread_mutex_t *m);
-
-TaskImpl *get_new_task();
 void allow_access_to_all_gpu_agents(void *ptr);
 }  // namespace core