--- a/openmp/libomptarget/plugins/hsa/impl/rt.h
+++ b/openmp/libomptarget/plugins/hsa/impl/rt.h
@@ -81,21 +81,6 @@ class Runtime {
   // modules
   atmi_status_t RegisterModuleFromMemory(void *, size_t,
                                          atmi_place_t);
-  // kernels
-  virtual atmi_status_t CreateKernel(atmi_kernel_t *, const int, const size_t *,
-                                     const char *);
-  virtual atmi_status_t ReleaseKernel(atmi_kernel_t);
-  atmi_status_t CreateEmptyKernel(atmi_kernel_t *, const int, const size_t *);
-  atmi_status_t AddGPUKernelImpl(atmi_kernel_t, const char *,
-                                 const unsigned int);
-  // sync
-  atmi_status_t TaskWait(atmi_task_handle_t);
-  // tasks
-  atmi_task_handle_t LaunchTask(atmi_lparm_t *, atmi_kernel_t, void **);
-  // taskgroups
-  atmi_status_t TaskGroupCreate(atmi_taskgroup_handle_t *, bool ordered = false,
-                                atmi_place_t place = ATMI_DEFAULT_PLACE);
-  atmi_status_t TaskGroupRelease(atmi_taskgroup_handle_t);
   // data
   atmi_status_t Memcpy(void *, const void *, size_t);
   atmi_status_t Memfree(void *);