… openmp/libomptarget/plugins/hsa/impl/atmi.h
… openmp/libomptarget/plugins/hsa/impl/atmi_runtime.h
… openmp/libomptarget/plugins/hsa/impl/kernel.cpp
… openmp/libomptarget/plugins/hsa/impl/kernel.h
… openmp/libomptarget/plugins/hsa/impl/machine.tcc
… openmp/libomptarget/plugins/hsa/impl/queue.cpp
… openmp/libomptarget/plugins/hsa/impl/queue.h
… openmp/libomptarget/plugins/hsa/impl/task.cpp
… openmp/libomptarget/plugins/hsa/impl/task.h
… openmp/libomptarget/plugins/hsa/impl/taskgroup.cpp
… openmp/libomptarget/plugins/hsa/impl/taskgroup.h