Prior to this patch, the local ATMI instance manages queues, kernarg memory, signals and kernel launch.

This patch moves that functionality into rtl.cpp and removes resulting dead code. It builds the same caching structure as ATMI for kernarg memory and signals. Kernel launch sets the same kernel packet values and initializes the same implicit args as before.

Functional differences:

ATMI allocates eight HSA queues per GPU and round robin schedules across them. I don't think that's a good fit for OpenMP as synchronising between kernels is made much more difficult by spreading kernels across multiple queues.

Uses a hash table instead of mapping kernels onto an integer ID.

Drops the (partial?) support for kernels with the same name and different argument count as I don't think we can do anything useful with that from OpenMP.

Performance improvement. Looks significant and reproducible, e.g. lulesh from 40s to 35s, snap from 40 to 30s. xsbench didn't change. My theory is that launching kernels directly reduced the latency.

I think there is a race somewhere in the interaction with rocr. Roughly 1% of runs of firstprivate2 report a corrupt aql packet. This change launches packets with the same backing memory and the same values as before (and no difference between passing and failing runs). That test runtime is reduced from 1.4 to 1.3s (repeatably), I think this increased the probability of hitting an existing fault.

Further improvement is possible - e.g. kernarg memory should probably be per-gpu, not per-kernel - but is left out of this patch to minimise the moving pieces.